M3-T24: the rough estimate stops truncating its own radius
B-063, ADR-0035. `radius_px = int(np.sqrt(median_area / np.pi))` was a second,
undeclared rounding — downward — three lines above the declared one that rounds
up.

Not an operator decision: the rule had been settled three times before this task
existed. ADR-0020 made `_integer_radius` the one funnel and the direction up;
ADR-0024 deleted this exact `int()` as D-04's mechanism; and the parameter's own
docstring, from the March 2026 commit that introduced both lines, calls
`scale=1.7` a multiplier making the disk "safely larger than a particle" — the
truncation made it smaller, from the day they were written together.

Delta: 730 differences across four phantoms. Large in count, small in magnitude:
320 under 1%, mean measured height <= 0.09%, and the largest scientifically
meaningful move is 2.5% at one phantom's 90th height percentile. The 21 moves
over 20% are baseline percentiles near zero.

The plan predicted ~70x less and the correction is the finding. Both of its
specific predictions held — the substrate moves on one phantom, the heights on
one — but it modelled only what `build_substrate_map` returns. `sizes` also
feeds `estimate_log_params`, so the LoG sigma range shifts wherever the rough
radius did: 379 of the differences are `log_detection`, and
afm_dense_overlapping detects one more particle (59 -> 60) with a byte-identical
substrate, which isolates the mechanism. The lesson is about the two-stage
design rather than this defect — the second stage is robust to the first, but
the diagnostics the first emits are wired straight into the detector and nothing
says so.

Detection quality, answerable for the first time thanks to M3-T15: recall
unchanged on every phantom. On the one whose substrate moved, radius error
improves 6% (0.765 -> 0.718 px, signed -0.704 -> -0.669) and localisation
degrades 0.3% in the mean while improving in the median. A wash, reported as
one. The signed error moving toward zero is consistent with a larger opening
leaving more of each particle above the substrate, but 6% on one phantom is not
evidence and no claim is made.

Filed, not fixed: B-064, the provenance of 1.7 and 2.5. Neither is derived
anywhere and both were tuned against a systematically small estimate — the
effective margin was 1.7 * int(r)/r, which at r = 4.9 is 1.39. M3-T15 can now
score a candidate, which makes it a task rather than a preference.

18 tests; restoring `int()` turns 11 red. One of them caught an overstatement in
its own first draft: at an equivalent radius of 3.432 truncation and the correct
value are both 6, and that case is now a named test, because it is also why the
defect was hard to see.

make check green, 452 tests. mypy unchanged at 12.

+++ nanoscope/core/science/preprocessing/substrate.py
@@ -224,9 +224,13 @@ def estimate_rough_radius(
         )
         return _fallback_radius(z, min_size_px)
 
-    # Median area -> equivalent radius
+    # Median area -> equivalent radius. **Not** truncated: `_integer_radius` is
+    # the one rounding a radius gets (ADR-0020), and this used to be a second,
+    # undeclared one going the other way — downward, ahead of the `* scale` that
+    # amplifies it, inside a parameter whose documented purpose is a margin
+    # making the disk *larger* than a particle (B-063, ADR-0035).
     median_area = np.median([p.area for p in props])
-    radius_px = int(np.sqrt(median_area / np.pi))
+    radius_px = float(np.sqrt(median_area / np.pi))
 
     # Scale up so the disk is safely larger than a particle
     rough_radius = max(radius_px * scale, min_size_px)
